requirements.txt
# LLM
openai>=1.50.0

# RAG - Embeddings & Vector Store
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Agentic AI
pyautogen==0.2.35

# MCP
mcp>=1.0.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
